chore: Implement 'nitpick' suggestions
