+++ bciers/libs/components/src/datagrid/createColumnGrid.tsx
@@ -1,0 +1,12 @@
+const createColumnGroup = (
+    groupId: string,
+    headerName: string,
+    renderHeaderGroup: any,
+) => ({
+    groupId,
+    headerName,
+    renderHeaderGroup,
+    children: [{ field: groupId }],
+});
+
+export default createColumnGroup;
